Add support for `import` statements
[in lib/cli.js]
 - Copying of IO file is no more required

[in lib/parser.js]
 - Call `importParser` in `statementParser`

[in lib/import]
 - Put IO class in `lib/import/core.js`
 - Separate `node` specific methods to `lib/import/node-core.js`
