+++ lib/cli.js
@@ -38,9 +38,7 @@ const cliEval = (argsObj) => {
     } else {
       tree.body = typeInferenceDisabled ? tree.body : inferTypes(tree.body)
       const out = escodegen.generate(tree, {comment: true})
-      const importIO = fs.readFileSync(path.join(__dirname, '/import/io-node.js'), 'utf8').toString()
-      const gen = importIO + '\n' + out
-      fs.writeFileSync(outfile, gen, 'utf8')
+      fs.writeFileSync(outfile, out, 'utf8')
       if (!outFilePath) {
         const mayBeCliArgs = commandLineArgs.length > 0 ? ' ' + commandLineArgs.join(' ') : ''
         childProcess.execSync('node ' + outfile + mayBeCliArgs, {stdio: 'inherit'})
