+++ lib/parser.js
@@ -28,7 +28,7 @@ const {
   ifParser, thenParser, elseParser,
   slashParser, parenCheck,
   reverseBindParser, doParser, ioFuncNameParser, ioMethodNameParser, //, assignmentOperatorParser
-  returnKeywordParser
+  returnKeywordParser, importParser
 } = base
 
 const valueParser = input => parser.any(objectParser, binaryExprParser, unaryExprParser, letExpressionParser,
@@ -527,7 +527,7 @@ const ifExprParser = input => mayBe(
     }
 )
 
-const statementParser = input => parser.any(multiLineCommentParser, singleLineCommentParser, returnParser, doBlockParser, ioParser, declParser, ifExprParser, fnDeclParser, fnCallParser, lambdaParser, lambdaCallParser)(input)
+const statementParser = input => parser.any(importParser, multiLineCommentParser, singleLineCommentParser, returnParser, doBlockParser, ioParser, declParser, ifExprParser, fnDeclParser, fnCallParser, lambdaParser, lambdaCallParser)(input)
 
 const programParser = (input, ast = estemplate.ast()) => {
   let [, rest] = returnRest('', input, input.str)
